requirements.txt
tensorflow-gpu
numpy
sklearn
# attrs